@@ -1,10 +1,8 @@
 <Project Sdk="Microsoft.NET.Sdk">
-  <ItemGroup>
-    <ProjectReference Include="..\NCoreUtils.Extensions.Collections.Abstractions\NCoreUtils.Extensions.Collections.Abstractions.csproj" />
-  </ItemGroup>
   <PropertyGroup>
-    <TargetFrameworks>net7.0;net6.0;netstandard2.1;net452;netstandard2.0</TargetFrameworks>
-    <Version>3.0.0</Version>
+    <TargetFrameworks>net8.0;net7.0;net6.0;netstandard2.1;net452;netstandard2.0</TargetFrameworks>
+    <Version>8.0.0</Version>
+    <IsAotCompatible Condition="$([MSBuild]::IsTargetFrameworkCompatible('$(TargetFramework)', 'net7.0'))">true</IsAotCompatible>
     <!-- NuGet -->
     <PackageId>NCoreUtils.Extensions.Collections</PackageId>
     <Title>$(PackageId)</Title>
@@ -24,9 +22,11 @@
     <GenerateDocumentationFile>true</GenerateDocumentationFile>
     -->
     <TreatWarningsAsErrors>true</TreatWarningsAsErrors>
-    <NoWarn>CS8618;CS8601;CS8625</NoWarn>
   </PropertyGroup>
-  <ItemGroup Condition=" '$(TargetFramework)' == 'net6.0' Or '$(TargetFramework)' == 'net7.0' ">
+  <ItemGroup>
+    <ProjectReference Include="..\NCoreUtils.Extensions.Collections.Abstractions\NCoreUtils.Extensions.Collections.Abstractions.csproj" />
+  </ItemGroup>
+  <ItemGroup Condition="$([MSBuild]::IsTargetFrameworkCompatible('$(TargetFramework)', 'net6.0'))">
     <Compile Remove="CodeAnalysis\**" />
   </ItemGroup>
   <ItemGroup Condition=" $(TargetFramework.StartsWith('net4')) Or '$(TargetFramework)' == 'netstandard2.0' ">
@@ -36,12 +36,4 @@
       <IncludeAssets>runtime; build; native; contentfiles; analyzers</IncludeAssets>
     </PackageReference>
   </ItemGroup>
-  <!-- mono fix -->
-  <PropertyGroup>
-    <TargetIsMono Condition="$(TargetFramework.StartsWith('net4')) and '$(OS)' == 'Unix'">true</TargetIsMono>
-    <BaseFrameworkPathOverrideForMono Condition="'$(BaseFrameworkPathOverrideForMono)' == '' AND '$(TargetIsMono)' == 'true' AND EXISTS('/usr/lib/mono')">/usr/lib/mono</BaseFrameworkPathOverrideForMono>
-    <FrameworkPathOverride Condition="'$(BaseFrameworkPathOverrideForMono)' != '' AND '$(TargetFramework)' == 'net452'">$(BaseFrameworkPathOverrideForMono)/4.5.2-api</FrameworkPathOverride>
-    <EnableFrameworkPathOverride Condition="'$(BaseFrameworkPathOverrideForMono)' != ''">true</EnableFrameworkPathOverride>
-    <AssemblySearchPaths Condition="'$(BaseFrameworkPathOverrideForMono)' != ''">$(FrameworkPathOverride)/Facades;$(AssemblySearchPaths)</AssemblySearchPaths>
-  </PropertyGroup>
 </Project>
