@@ -2,15 +2,14 @@ using System.Diagnostics;
 using System.Runtime.CompilerServices;
 using System.Runtime.Serialization;
 
-namespace NCoreUtils.Collections
+namespace NCoreUtils.Collections;
+
+internal static class SerializationInfoExtensions
 {
-    static class SerializationInfoExtensions
+    [MethodImpl(MethodImplOptions.AggressiveInlining)]
+    [DebuggerStepThrough]
+    public static T GetValue<T>(this SerializationInfo info, string name)
     {
-        [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        [DebuggerStepThrough]
-        public static T GetValue<T>(this SerializationInfo info, string name)
-        {
-            return (T)info.GetValue(name, typeof(T))!;
-        }
+        return (T)info.GetValue(name, typeof(T))!;
     }
 }
