@@ -3,61 +3,63 @@ using System.Collections.Generic;
 using System.Diagnostics.CodeAnalysis;
 using System.Runtime.CompilerServices;
 
-namespace NCoreUtils
+namespace NCoreUtils;
+
+/// <summary>
+/// Defines dictionary extension methods.
+/// </summary>
+public static class DictionaryExtensions
 {
     /// <summary>
-    /// Defines dictionary extension methods.
+    /// Either returns value associated with the specified key or the default value.
     /// </summary>
-    public static class DictionaryExtensions
+    /// <param name="dictionary">Source dictionary.</param>
+    /// <param name="key">Key to try.</param>
+    /// <param name="defaultValue">Value returned if no value associated with the specified key.</param>
+    /// <returns>Either value associated with the specified key or the default value.</returns>
+    [MethodImpl(MethodImplOptions.AggressiveInlining)]
+    [return: NotNullIfNotNull(nameof(defaultValue))]
+#if NET6_0_OR_GREATER
+    [Obsolete("Use GetValueOrDefault instead")]
+#endif
+    public static TValue? GetOrDefault<TKey, TValue>(
+        this IReadOnlyDictionary<TKey, TValue> dictionary,
+        TKey key,
+        TValue? defaultValue = default)
     {
-        /// <summary>
-        /// Either returns value associated with the specified key or the default value.
-        /// </summary>
-        /// <param name="dictionary">Source dictionary.</param>
-        /// <param name="key">Key to try.</param>
-        /// <param name="defaultValue">Value returned if no value associated with the specified key.</param>
-        /// <returns>Either value associated with the specified key or the default value.</returns>
-        [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        [return: NotNullIfNotNull("defaultValue")]
-        public static TValue? GetOrDefault<TKey, TValue>(
-            this IReadOnlyDictionary<TKey, TValue> dictionary,
-            TKey key,
-            TValue? defaultValue = default)
+        if (dictionary == null)
         {
-            if (dictionary == null)
-            {
-                throw new ArgumentNullException(nameof(dictionary));
-            }
-            if (dictionary.TryGetValue(key, out var value))
-            {
-                return value;
-            }
-            return defaultValue;
+            throw new ArgumentNullException(nameof(dictionary));
         }
+        if (dictionary.TryGetValue(key, out var value))
+        {
+            return value;
+        }
+        return defaultValue;
+    }
 
-        /// <summary>
-        /// Returns the value associated with the specified key, or a default value if the dictionary has no value
-        /// assiciated to the key.
-        /// </summary>
-        /// <param name="source">Source dictionary.</param>
-        /// <param name="key">The key.</param>
-        /// <param name="valueFactory">Value factory invoked to provide default value.</param>
-        /// <returns>
-        /// Either the value associated with the specified key, or a default value if the dictionary has no value
-        /// assiciated to the key.
-        /// </returns>
-        [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        public static TValue GetOrSupply<TKey, TValue>(this IReadOnlyDictionary<TKey, TValue> source, TKey key, Func<TValue> valueFactory)
+    /// <summary>
+    /// Returns the value associated with the specified key, or a default value if the dictionary has no value
+    /// assiciated to the key.
+    /// </summary>
+    /// <param name="source">Source dictionary.</param>
+    /// <param name="key">The key.</param>
+    /// <param name="valueFactory">Value factory invoked to provide default value.</param>
+    /// <returns>
+    /// Either the value associated with the specified key, or a default value if the dictionary has no value
+    /// assiciated to the key.
+    /// </returns>
+    [MethodImpl(MethodImplOptions.AggressiveInlining)]
+    public static TValue GetOrSupply<TKey, TValue>(this IReadOnlyDictionary<TKey, TValue> source, TKey key, Func<TValue> valueFactory)
+    {
+        if (source == null)
         {
-            if (source == null)
-            {
-                throw new ArgumentNullException(nameof(source));
-            }
-            if (valueFactory == null)
-            {
-                throw new ArgumentNullException(nameof(valueFactory));
-            }
-            return source.TryGetValue(key, out var value) ? value : valueFactory();
+            throw new ArgumentNullException(nameof(source));
         }
+        if (valueFactory == null)
+        {
+            throw new ArgumentNullException(nameof(valueFactory));
+        }
+        return source.TryGetValue(key, out var value) ? value : valueFactory();
     }
 }
